+ Add ECDH and ECC keygen / import perf tests for CNG to validate changes in BCrypt in this area
+ Add SymCryptUnitTest option to skip perf tests based on key size so
  that implementations don't need to support all or nothing vs. SymCrypt
  + Apply to new CNG perf tests, and similar OpenSSL perf tests that
    were reporting bogus perf for unsupported curves

Related work items: #60270674

+++ lib/parhash.c
@@ -457,7 +457,7 @@ SymCryptParallelHashProcess(
             todo = SYMCRYPT_MIN( todo, pWork[i]->cbData );
         }
 
-        nBytes = todo & ~(pHash->inputBlockSize - 1 );
+        nBytes = todo & ~((SIZE_T)(pHash->inputBlockSize - 1));
 
         (*pParHash->parAppendFunc)( pWork, nPar, nBytes, pbFixedScratch, cbFixedScratch );
